refactor: move Lambda code inside `src`, extract App config
