+++ layers/mathjax-node-layer/.gitignore
