+++ layers/mathjax-node-layer/nodejs/package.json
